backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary # Driver PostgreSQL
python-dotenv
Flask-Cors